requirements.txt
# ============================================================
# JobPredator — Agentic Job Hunter
# ============================================================

# --- Core async & HTTP ---
httpx>=0.27.0
aiohttp>=3.9.0
requests>=2.31.0
tenacity>=8.2.0

# --- Job scraping ---
python-jobspy>=1.1.80
playwright>=1.44.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
trafilatura>=1.12.0          # semantic HTML extraction
playwright-stealth>=1.0.0    # anti-bot detection
PyGithub>=2.3.0              # GitHub API client

# --- CV parsing ---
pymupdf>=1.24.0             # aka fitz — fastest PDF parser
pdfplumber>=0.11.0          # best for structured/table PDFs
python-docx>=1.1.0          # DOCX read/write
pylatexenc>=2.10            # Overleaf LaTeX parsing
pytesseract>=0.3.10         # OCR fallback for scanned PDFs
Pillow>=10.0.0              # image support for OCR

# --- LLM & AI ---
openai>=1.30.0              # Azure OpenAI compatible client
langchain>=0.2.0
langchain-openai>=0.1.8
langchain-community>=0.2.0
langgraph>=0.1.5
langchain-core>=0.2.0
sentence-transformers>=3.0.0  # local embeddings (consider BAAI/bge-base-en-v1.5)
tiktoken>=0.7.0
instructor>=1.2.0           # structured LLM outputs
networkx>=3.3               # knowledge graph support
scikit-learn>=1.5.0         # ML utilities for ranking

# --- Database ---
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0             # async PostgreSQL driver
alembic>=1.13.0
pgvector>=0.3.0             # pgvector Python client

# --- Data ---
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# --- Cover letter & document export ---
jinja2>=3.1.4
weasyprint>=62.0            # PDF export from HTML
reportlab>=4.2.0            # alternative PDF generation

# --- Email & outreach ---
# (uses stdlib smtplib + aiosmtplib for async)
aiosmtplib>=3.0.0
email-validator>=2.1.0

# --- API & CLI ---
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
typer>=0.12.0
rich>=13.7.0                # beautiful CLI output

# --- Config & utils ---
python-dotenv>=1.0.1
pyyaml>=6.0.1
loguru>=0.7.2
tqdm>=4.66.0

# --- Dev & testing ---
pytest>=8.2.0
pytest-asyncio>=0.23.0
black>=24.4.0
